Back to 16

+++ apps/api/src/utils/helpers.ts
@@ -69,7 +69,7 @@ async function loadClip(): Promise<{
       "Xenova/clip-vit-large-patch14-336",
       {
         device: "cpu",
-        dtype: "fp32",
+        dtype: "fp16",
       }
     );
     clipProcessorPromise = AutoProcessor.from_pretrained(
